requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pybit==5.8.0
pycryptodome==3.21.0
python-telegram-bot==21.6
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.3
websocket-client==1.8.0
Flask
gunicorn==20.1.0
python-dotenv==1.0.1
quart
hypercorn
python-telegram-bot[webhooks]